# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Periscol Backend API documentation
  version: 0.0.1
  title: Periscol Backend API
  contact: {}
  license:
    name: unlicensed
host: localhost:8080
basePath: /
...
paths:
...
  /api/child:
    post:
      tags:
        - child-resource
      summary: createChild
      operationId: createChildUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: childDTO
          description: childDTO
          required: true
          schema:
            $ref: '#/definitions/ChildDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ChildDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/child/{id}:
    get:
      tags:
        - child-resource
      summary: getChild
      operationId: getChildUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ChildDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - child-resource
      summary: updateChild
      operationId: updateChildUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: childDTO
          description: childDTO
          required: true
          schema:
            $ref: '#/definitions/ChildDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ChildDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - child-resource
      summary: deleteChild
      operationId: deleteChildUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - child-resource
      summary: partialUpdateChild
      operationId: partialUpdateChildUsingPATCH
      consumes:
        - application/json
        - application/merge-patch+json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: childDTO
          description: childDTO
          required: true
          schema:
            $ref: '#/definitions/ChildDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ChildDTO'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/children:
    get:
      tags:
        - child-resource
      summary: getAllChildren
      operationId: getAllChildrenUsingGET
      produces:
        - '*/*'
      parameters:
        - name: eagerload
          in: query
          description: eagerload
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ChildDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/diet:
    post:
      tags:
        - diet-resource
      summary: createDiet
      operationId: createDietUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dietDTO
          description: dietDTO
          required: true
          schema:
            $ref: '#/definitions/DietDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DietDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/diet/{id}:
    get:
      tags:
        - diet-resource
      summary: getDiet
      operationId: getDietUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DietDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - diet-resource
      summary: updateDiet
      operationId: updateDietUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dietDTO
          description: dietDTO
          required: true
          schema:
            $ref: '#/definitions/DietDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DietDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - diet-resource
      summary: deleteDiet
      operationId: deleteDietUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - diet-resource
      summary: partialUpdateDiet
      operationId: partialUpdateDietUsingPATCH
      consumes:
        - application/json
        - application/merge-patch+json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dietDTO
          description: dietDTO
          required: true
          schema:
            $ref: '#/definitions/DietDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DietDTO'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/diets:
    get:
      tags:
        - diet-resource
      summary: getAllDiets
      operationId: getAllDietsUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/DietDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
...
  ChildDTO:
    type: object
    properties:
      adelphie:
        $ref: '#/definitions/FamilyDTO'
      birthday:
        type: string
        format: date-time
      classroom:
        $ref: '#/definitions/ClassroomDTO'
      diets:
        type: array
        uniqueItems: true
        items:
          $ref: '#/definitions/DietDTO'
      gradeLevel:
        $ref: '#/definitions/GradeLevelDTO'
      id:
        type: integer
        format: int64
      name:
        type: string
      surname:
        type: string
    title: ChildDTO
  ClassroomDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      professor:
        type: string
    title: ClassroomDTO
  DietDTO:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
    title: DietDTO
  FamilyDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      postalAdress:
        type: string
      referingParentName:
        type: string
      referingParentSurname:
        type: string
      telephoneNumber:
        type: string
    title: FamilyDTO
  GradeLevelDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      level:
        type: string
    title: GradeLevelDTO
...